+++ data-raw/metadata.R
@@ -33,7 +33,8 @@ metadata <- tibble::tribble(
   "Instituciones deportivas", "IDE", "MIDES", 32721, "wfs", 2015, "https://mapas.mides.gub.uy/geoserver/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=IDE:inst_deportivas_2015", "gml_id", "nombre",
   "Postes Kilometros", "MTOP", "MTOP", 32721, "wfs", 2019, "https://geoservicios.mtop.gub.uy/geoserver/inf_tte_ttelog_terrestre/postes_km/ows?service=WFS&request=GetFeature&typeName=postes_km", NA, NA,
   "Grilla ortofotos urbana", "IDE", "IDE", 5381, "wfs", 2019, "https://mapas.ide.uy/geoserver-raster/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=ortofotos:grilla_urbana", "gml_id","nombre",
-  "Grilla ortofotos nacional", "IDE", "IDE", 5381, "wfs", 2019, "https://mapas.ide.uy/geoserver-raster/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=ortofotos:grilla_nacional", "gml_id","nombre"
+  "Grilla ortofotos nacional", "IDE", "IDE", 5381, "wfs", 2019, "https://mapas.ide.uy/geoserver-raster/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=ortofotos:grilla_nacional", "gml_id","nombre",
+  "Ambientes acuaticos", "DINAMA", "MVOTMA", 32721, "zip", 2019, "https://www.dinama.gub.uy/geoserver/u19600217/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=u19600217:c094&outputFormat=SHAPE-ZIP", NA, NA
 )
 
 # saveRDS(metadata,"data-raw/metadata.Rds")